Auto-Selection of next talk in a room at an event
Adds a feature to allow Freeseer to automatically detect the next talk
based on time window within a 15 minute window starting from the current
time.

This event is triggered when the user clicks the STOP button at the end
of a recording. This code uses the system's local time +15 minutes to
query the database and picks the first talk returned.

Closes #65