Added checkbox to flag if a release form is received
This was implemented in the Report Dialog as well as Report Editor.
Since we are flagging Release Forms now adding a new failure type of "No
Issues" was necessary to flag talks where there was nothing wrong but a
release was received.

As a note in the DB Sqlite does not support boolean types so the value
for if a report is received or not is represented as integers 0-False,
1-True.

Closes #174