Close #193 - Can't record 2x in a row
This is a simple workaround which ignores the TypeError that is thrown when
recording the 2nd time in a row with an empty plugin config.

This issue is triggered by an empty config for the plugin when recording a 2nd
time in a row. The plugin will attempt to read the non-existent setting from
the config file and throw a TypeError. A manual user workaround is to have the
user toggle the videopassthrough plugin configuration screen in freeseer-config
which will trigger the plugin to save it's settings to file.

This code change ignores the TypeError which doesn't seem to have any adverse
effects on using Freeseer.