Implemented RTMP streaming
NOT YET TESTED.  (this line will be removed from commit message when testing is complete)

Implemented get_output_bin function in RTMP Streaming plugin. Returns
a bin with video and audio sink pads, as necessary.  Uses flvmux to mux the audio/video streams, and add
metadata tags.  Uses x264enc to encode the video to x-h264 for flv.  Streams flv content to rtmp server at url provided by user.

Testing notes: TODO

Notes for review:
- I add the metadata tags to the flv stream using the flvmux element.
Currently, I'm not sure if that's the right way how to do it, and
need to look into how to verify it.
- I set the bitrate for the h264 encoder based on the value found
in the ogg-output plugin.