GH-206 Fixes to RTMP Plugin
Corrected changes to rtmp streaming, so that changes made using
the config widget take effect.  Also, commented out audio encoding
lines, until faac becomes available, or we decideo to use another
encoder or element to adjust audio quality.

(most of these need for fixes and the weird commit patterns were a result
of me doing development on my local machine, and testing on my vm,
and using git to transfer between the two.  I have since decided that
was a not-the-most-intelligent way of going about things.)