Began work on Youtube live events manager
Started work on youtube live events manager, intented to
send api requests to the server, and handle the response.  Currently
(though none of it has been tested), it builds and sends requests
to the server for api calls for managing live events.  It does
not parse the response yet, or do anything with it at all right now.

The intended design would have had some sort of interface for
creating and managing the live event in the talk editor perhaps,
which would create, update, start, end, or delete the events.  Core
would then have to set the url of the RTMP Streaming plugin based on
the url received in the response.  As well, CDN (encoding format) data
may have to be included in the api calls (but no testing has been done regarding this)