Remove "Technical Writer" page
Not necessary. The readme of the freeseer-docs repo does a good job
introducing the documentation tools and how to get started.

Note that the documentation.rst page also mentioned sending in
documentation patches by email (for those who weren't comfortable
with Git or GitHub). I'm sure we'll still allow this, but it's never
happened so far, so we can safely remove the mention of it.