Add ToC to top of Basics page
The Basics page has a lot of content on it and the sections are all
independent so scrolling through the page to find something doesn't
work well. The ToC on the sidebar is difficult to read when titles
are long and it's not obvious, so I'm putting a more obvious ToC
on the page to make navigation easier.