In flight migration to GStreamer 1.0
- This commit is pretty much so that the project can see this migration
in progress. It will is all likeliness be squashed later.

- Having limited the functionality of this migration to; working in
Linux, and Windows using the test source, I am starting to make so
actual progress.

- Up to this point the project builds without errors, in Windows, and if
you select the test source as the source it will play. Unfortunately
currently GStreamer opens its own window for the preview instead of
painting it in FreeSeer

- I would pretty much assume this will fail Travis