Changed the color format box
- As per the comments related to this feature I have changed the drop
down box in the GUI to give you the option of selecting the, Y444, RGBx,
or YVYU, color formats. It does not make much sense to have a drop box
with two options that are the same.

- I had originally selected Y444 because it was the format used in an
example from Novacut
(http://bazaar.launchpad.net/~jderose/+junk/gst-examples/view/head:/transcoder-1.0)

- There is a long list of possible video formats that I found here:

http://gstreamer.freedesktop.org/data/doc/gstreamer/head/gst-plugins-base-plugins/html/gst-plugins-base-plugins-videotestsrc.html

- It is a simple process to change the color format and if the desired
format is not implemented it just needs to be added in widget.py

- The 3 I have coded work with the test source in Windows