Fix Schema Upgrade code
Schemas are now versioned and used when creating new tables. The table
create methods itself will default to the latest schema versions so they
can be used without code changes.

If the presentations table is being creating for the first time. The
upgrade function will no longer need to run. This is achieved by setting
the DB version immediately.

Fixed the default talk to not be blank.

Change-Id: Ibdf9b9b28156ef4dafeefe73ef27124fb0cd5e57