Remove Eclipse project file
Was used for the freeseer-docs repository so Eclipse users did not have
to configure a new Eclipse project just to edit the docs.

If we do want to have an .project file, it should be moved to the
project root. And then we may as well use it for more stuff like
settings tabs to spaces, configuring the run button, and so on.