Added AboutWidget
Added AboutWidget to the left hand tree on the config tool. This
widget is based off the previous AboutDialog. It contains information
about Freeseer, including copyright and version. There are also three
buttons, one links the user to the freeseer read the docs page
for help, the second other lets them report an issue on github and
the third button links to the contact page on read the docs.

The AboutDialog has also been modified. Essentially it was converted
to a widget (AboutWidget), which is used now for both AboutDialog and
the config tool.

Related to #411