Redesign of the Settings List
Make the interface a lot less complicated and confusing. There
are no configuration options in the plugins anymore. It is
simply a list of all the plugins installed. The list starts
fully expanded, but sections may be minimized. When a plugin
is clicked, further details are displayed under the list.

Close #535
Related to issue #411