Revert "Close #496 Set travis git clone depth to 1"
The problem with using git clone depth 1 is that if the contributer
pushes multiple commits in short succession before Travis is able to
kick off a build. The build will fail due to being unable to find the
commit. This is because Travis checks out a branch and then checks out
the exact commit. So it is possible to get into a situation where the
latest branch is newer than the commit Travis is trying to build
against.

This reverts commit 217a887c2341ffa4bf12e7fc29bbf6335436ae5d.