Fix #592 repaired documentation cross-references
The about page and contribute basics page contained sphinx cross
references that did not point to valid files. The cross references have
been updated to point to new pages. The new cross references point to
similarily named directories to try and preserve the original intent of
the cross references.

Fix #592