Fix #592 added warning to quick-start guide
The problem that caused import errors as in #592 were caused by a pip
installation of Freeseer and a developer working with a source copy on
the same machine. This commit places a warning at the beginning of the
quick-start page that tells readers to use the developer installation
if they want to modify the source code.

Fix #592