Add Instructions to Installation for Development
An introductory paragraph has been added to the "Installing Freeseer
for Development" section, giving an overall summary of how to install
Freeseer for development. It links to the guide on how to fork and
clone the project at the beginning of the section instead of the end.

In addition, the dollar signs $ have been removed from any command
line instructions for consistency and clarity.

Related to #442
Close #599